GaussianDispersion

Documentation for GaussianDispersion.

GaussianDispersion.DispersionParams

GaussianDispersion.GaussianDispersionParams

GaussianDispersion.GaussianDispersionParams

GaussianDispersion.HeatBalanceParams

GaussianDispersion.PasquillGiffordCriteria

GaussianDispersion.ReleaseParams

GaussianDispersion.Stabilities

GaussianDispersion.albedo_elevation

GaussianDispersion.briggs_dispersion

GaussianDispersion.buoyancy_flux

GaussianDispersion.disp_function

GaussianDispersion.friction_temp

GaussianDispersion.friction_velocity

GaussianDispersion.height_hs

GaussianDispersion.height_isa

GaussianDispersion.hour_angle

GaussianDispersion.net_radiation

GaussianDispersion.obukhov

GaussianDispersion.pasquill_gifford

GaussianDispersion.plume_rise

GaussianDispersion.potential_temp

GaussianDispersion.saturation_pressure

GaussianDispersion.sensible_heat_flux

GaussianDispersion.solar_declination

GaussianDispersion.solar_elevation

GaussianDispersion.solar_energy_flux

GaussianDispersion.virtual_temp

GaussianDispersion.virtual_temp

GaussianDispersion.DispersionParams — Method

DispersionParams(x::Real, terrain::AbstractTerrain, stability::StabilityClass)

Return a DispersionParams struct given the downwind direction x, the terrain and stability class

Example

julia> DispersionParams(100, Rural, A) DispersionParams(21.89081818461976, 20.0) ```

GaussianDispersion.GaussianDispersionParams — Type

mutable struct GaussianDispersionParams

Structure related to the release conditions

release

release parameters

terrain

terrain (Urban/Rural)

stabilities

set of stability classes

reflection

if ground reflection is considered

hmix

height of the mixing layer

GaussianDispersion.GaussianDispersionParams — Method

Return the concentration in [g m^{-3}] at some point given the conditions stated in params.

x is the downwind distance from the point source.

y is the horizontal distance perpendicular to the wind.

z is the vertical distance from the point source.

GaussianDispersion.HeatBalanceParams — Type

mutable struct HeatBalanceParams

Values for the heat balance parametrization.

'albedo' :

C_g: parameter for the estimation of the heat flux to the ground in the equation $q\_G = C\_G R\_n$.

Typical values are 0.05-0.25 for rural, 0.25-0.3 for urban and 0.1 for grasslands (Scire et al., 2000)

GaussianDispersion.PasquillGiffordCriteria — Type

PasquillGiffordCriteria

Critera for determining the stability class according to sky conditions (Pasquill, 1961; Gifford, 1961) Strong, Moderate and Slight are for incoming solar radiation (for daytime) Cloudy and Clear are for cloudiness (for nighttime)

GaussianDispersion.ReleaseParams — Type

mutable struct ReleaseParams

Structure related to the release conditions

h

Effective source height [m]

Q

Emission rate [g/s]

u

Wind speed [m/s]

GaussianDispersion.Stabilities — Method

Stabilities(
    stabs::Union{String, Symbol}...
) -> Set{StabilityClass}

Define the stability classes for the model according to Pasquill and Gifford (Pasquill, 1961; Gifford, 1961). If multiple classes are defined, the average result for each class is considered.

Example

julia> Stabilities(:A, :B)
Set{StabilityClass} with 2 elements:
  GaussianDispersion.A
  GaussianDispersion.B

GaussianDispersion.albedo_elevation — Method

albedo_elevation(albedo_90, ϕ) -> Any

Albedo given albedo_90 the albedo at 90° and ϕ the solar elevation

GaussianDispersion.briggs_dispersion — Method

briggs_dispersion(terrain::AbstractTerrain, stability::StabilityClass)

Given the terrain and the stability, return the three coefficients a, b, c for σy and σz respectively.

Example

julia> briggs_dispersion(Rural, A)
((a = 0.22, b = 0.0001, c = -0.5), (a = 0.2, b = 0, c = 1.0))

GaussianDispersion.buoyancy_flux — Method

buoyancy_flux(
    gas_density,
    air_density,
    stack_radius,
    gas_velocity
)

Buoyancy flux parameter according to Briggs (1968).

GaussianDispersion.disp_function — Method

disp_function(a, b, c)

Return the function for calculating the dispersion parameters with respect to x given the coefficients of the equation ax(1 + bx)^c ```

GaussianDispersion.friction_temp — Method

friction_temp(q, rho, u_star, c_p) -> Any

Calculate the friction temperature from the surface pressure p [Pa], temperature t [K], dewpoint temp td [K] and stress stress [N/m²].

Arguments

q: surface sensible heat flux [W/m²]

rho: air density [kg/m³]

u_star: scale velocity [m/s]

c_p: specific heat capacities of dry air [J/kg/K]

GaussianDispersion.friction_velocity — Method

friction_velocity(p, t, td, stress, R_gas) -> Any

Calculate the friction velocity from the surface pressure p [Pa], temperature t [K], dewpoint temp td [K] and stress stress [N/m²].

GaussianDispersion.height_hs — Method

height_hs(p, ps, R_gas, T, g) -> Any

Get height from pressure level in a hydrostatic atmosphere.

GaussianDispersion.height_isa — Method

height_isa(p, R_gas, g) -> Any

Get height from pressure level in the international standard atmosphere.

GaussianDispersion.hour_angle — Function

hour_angle(t) -> Any
hour_angle(t, t0) -> Any

t the time in hour and t0 the time of the solar noon.

GaussianDispersion.net_radiation — Method

net_radiation(R, albedo, cloud_cover, T_surf) -> Any

Net radiation energy flux received by the earth surface [W m^-2] with R the solar flux (solar_energy_flux)

GaussianDispersion.obukhov — Function

obukhov(ps, ts, td, t, p, u_star, q) -> Any
obukhov(ps, ts, td, t, p, u_star, q, c_p) -> Any
obukhov(ps, ts, td, t, p, u_star, q, c_p, R_gas) -> Any
obukhov(
    ps,
    ts,
    td,
    t,
    p,
    u_star,
    q,
    c_p,
    R_gas,
    karman
) -> Any
obukhov(
    ps,
    ts,
    td,
    t,
    p,
    u_star,
    q,
    c_p,
    R_gas,
    karman,
    g
) -> Any

Calculate Obukhov scale height from surface meteorological data and sensible heat flux.

Arguments

ps: surface pressure [Pa]

ts: surface temperature [K]

td: surface dew point [K]

t: temperature first model level [K]

p : pressure first model level [Pa]

u_star: scale velocity [m/s]

q: surface sensible heat flux [W/m2]

GaussianDispersion.pasquill_gifford — Method

pasquill_gifford(criteria::PasquillGiffordCriteria, windspeed::Real)

Return a list of stability classes according to the Pasquill Gifford criteria

Example

julia> pasquill_gifford(Moderate, 5.5) Set{StabilityClass} with 2 elements: C D

GaussianDispersion.plume_rise — Method

plume_rise(flux_param, x, u)

Plume rise Δh [m] according to Briggs (1975) parametrization. x is the downwind distance in meter and u the downwind velocity in meter/second.

GaussianDispersion.potential_temp — Method

potential_temp(t, p, akap) -> Any

Calculate the potential temperature.

Arguments

t: temperature [K]

p: pressure [Pa]

R_gas: individual gas constant for dry air [J/kg/K]

c_p: specific heat capacities of dry air [J/kg/K]

GaussianDispersion.saturation_pressure — Method

saturation_pressure(t) -> Any

Calculate the saturation vapor pressure from Arden Buck equations (Buck, 1996). t is in K, result in Pa.

GaussianDispersion.sensible_heat_flux — Method

sensible_heat_flux(C_g, net_rad, B) -> Any

Sensible heat flux to the surface [W m^-2]. net_rad can be calculated with net_radiation

GaussianDispersion.solar_declination — Method

solar_declination(day) -> Any

Latitude [°] where the sun is in the zenith at the solar noon. day is the day of the year

GaussianDispersion.solar_elevation — Method

solar_elevation(lat, sol_dec, h_angle) -> Any

Solar elevation [°] given the latitude, the solar declination and the hour angle.

GaussianDispersion.solar_energy_flux — Method

solar_energy_flux(Φ, n) -> Any

Solar radiation energy flux W m^-2 given the solar elevation Φ and n the fractional cloud cover. Ref: Kasten and Czeplak, 1980; Holtslag and van Ulden, 1983.

GaussianDispersion.virtual_temp — Method

virtual_temp(ps, t, e) -> Any

Calculate the virtual temperature from the surface pressure ps [Pa], the temperature t [K], and the vapor pressure e [Pa].

GaussianDispersion.virtual_temp — Method

virtual_temp(t, w) -> Any

Calculate the virtual temperature from the temperature t [K] and the specific humidity w [kg/kg].